% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbaxis.R
\name{ggbaxis}
\alias{ggbaxis}
\title{ggbaxis}
\usage{
ggbaxis(p, axis = "x", breaks)
}
\arguments{
\item{p}{ggplot object}

\item{axis}{one of 'x' or 'y'}

\item{breaks}{break interval}
}
\value{
ggplot object
}
\description{
prototype of using facet_grid to mimic axis break
}
\examples{
\dontrun{
d = data.frame(x=c(rnorm(5)+4,rnorm(5)+20,rnorm(5)+5,rnorm(5)+22), y=1:20)
mymin = function(x) ifelse(x <= 7, 0, 17)
p <- ggplot(d, aes(x, y)) + geom_rect(aes(xmin=mymin(x), xmax=x, ymin=y-.4, ymax=y+.4))
ggbaxis(p, axis="x", breaks=c(7, 17))
}
}
\author{
guangchuang yu
}
