% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcake.R
\name{ggcake}
\alias{ggcake}
\title{ggcake}
\usage{
ggcake(cake_color = "#FF3399")
}
\arguments{
\item{cake_color}{color of cake}
}
\value{
ggplot object
}
\description{
draw a cake using ggplot2
}
\references{
\url{https://www.r-bloggers.com/fun-with-r-graphics-a-raptor-and-a-cake/}
}
\author{
guangchuang yu
}
