% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpar.R
\name{ggpar}
\alias{ggpar}
\title{ggpar}
\usage{
ggpar(...)
}
\arguments{
\item{...}{parameters pass to par}
}
\description{
setting graphic parameter
}
\details{
setting graphic parameter via `par` for base plot and apply it to ggplot2. Currently only works for par(mar).
}
\author{
Guangchuang Yu
}
