% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggqrcode.R
\name{ggqrcode}
\alias{ggqrcode}
\title{ggqrcode}
\usage{
ggqrcode(text, color = "black", alpha = 1)
}
\arguments{
\item{text}{text string}

\item{color}{color}

\item{alpha}{[0, 1] for transparency}
}
\value{
ggplot object
}
\description{
generate qrcode by ggplot
}
\author{
guangchuang yu
}
