% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvenn.R
\name{ggvenn}
\alias{ggvenn}
\title{ggvenn}
\usage{
ggvenn(x, alpha = 0.5)
}
\arguments{
\item{x}{data}

\item{alpha}{transparency of color}
}
\value{
ggplot object
}
\description{
venn plot using ggplot2
}
\examples{
set.seed(2017-11-08)
x <- matrix(sample(0:4, 40, TRUE, c(.5, .1, .1, .1, .1)), ncol=4)
colnames(x) <- LETTERS[1:4]
ggvenn(x)
}
\author{
guangchuang yu
}
