% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv2file.R
\name{gv2file}
\alias{gv2file}
\title{gv2file}
\usage{
gv2file(gv, file, width = NULL, height = NULL)
}
\arguments{
\item{gv}{grViz object}

\item{file}{output file}

\item{width}{width}

\item{height}{height}
}
\value{
raw
}
\description{
convert gv (grViz object, output of DiagrammeR) to image file
}
\author{
Guangchuang
}
