% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meme.R
\name{meme_save}
\alias{meme_save}
\title{meme_save}
\usage{
meme_save(meme, file = "meme.png", device = NULL, dpi = 300)
}
\arguments{
\item{meme}{meme output}

\item{file}{output file}

\item{device}{graphic device}

\item{dpi}{dpi of the figure}
}
\description{
save meme plot
}
\author{
guangchuang yu
}
