% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaPrettyPlot.R
\name{msaPrettyPlot}
\alias{msaPrettyPlot}
\title{msaPrettyPlot}
\usage{
msaPrettyPlot(x, ..., density = 300)
}
\arguments{
\item{x}{an instance of MultipleAlignment}

\item{...}{additional parameters passed to msaPrettyPrint}

\item{density}{resolution for parsing temp PDF file}
}
\value{
ggplot object
}
\description{
msa plot
}
\details{
convert the msa::msaPrettyPrint output to a ggplot object
}
\author{
guangchuang yu
}
