% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_trend.R
\name{plot_citation_trend}
\alias{plot_citation_trend}
\title{plot_citation_trend}
\usage{
plot_citation_trend(id, pubid)
}
\arguments{
\item{id}{google scholar ID}

\item{pubid}{publication id}
}
\value{
ggplot object
}
\description{
plot citation trend
}
\author{
guangchuang yu
}
