% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dlstats_bioc.R
\name{plot_dlstats_bioc}
\alias{plot_dlstats_bioc}
\title{plot_dlstats_bioc}
\usage{
plot_dlstats_bioc(pkg)
}
\arguments{
\item{pkg}{package name}
}
\value{
ggplot object
}
\description{
plot download stats of Bioconductor package
}
\author{
guangchuang yu
}
