% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_trend.R
\name{pubmed_trend}
\alias{pubmed_trend}
\title{pubmed_trend}
\usage{
pubmed_trend(searchTerm, year, verbose = TRUE)
}
\arguments{
\item{searchTerm}{search term}

\item{year}{year vector}

\item{verbose}{whether print message}
}
\value{
data.frame
}
\description{
query pubmed record numbers of search term through years
}
\examples{
\dontrun{
pubmed_trend("Yu Guangchuang[Full Author Name]", 2010:2016)
}
}
\author{
guangchuang yu
}
