% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cake.R, R/geom_ord_ellipse.R, R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{geom_cake}
\alias{geom_ord_ellipse}
\alias{ggsave}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{gglayer}{\code{\link[gglayer]{geom_cake}}, \code{\link[gglayer]{geom_ord_ellipse}}}

  \item{ggplot2}{\code{\link[ggplot2]{ggsave}}}
}}

