% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_axis_text_color.R
\name{set_axis_text_color}
\alias{set_axis_text_color}
\title{set_axis_text_color}
\usage{
set_axis_text_color(p, colors, mapping, fake_aes = "color", axis = "x")
}
\arguments{
\item{p}{ggplot object}

\item{colors}{named vector of colors}

\item{mapping}{mapping variable for generating legend}

\item{fake_aes}{fake aes that use to add a invisible layer for generating legend}

\item{axis}{one of 'x', 'y', or 'xy'}
}
\value{
ggplot object
}
\description{
color axis text by specifying color and add legend based on mapping variable
}
\examples{
require(ggplot2)
set.seed(2017-07-04)
d <- data.frame(x = letters[1:5], y = rnorm(5),
	group = sample(c("Control", "Treatment"), 5, replace=TRUE),
	type = sample(LETTERS[1:2], 5, replace=TRUE)
)
cols <- c(A = "red", B = "blue")
p <- ggplot(d, aes(x,y)) + geom_col(aes(fill=group))
set_axis_text_color(p, colors=cols, mapping=~type, fake_aes='color')
}
\author{
guangchuang yu
}
