% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_font.R
\name{set_font}
\alias{set_font}
\title{set_font}
\usage{
set_font(p, family = "sans", fontface = NULL, size = NULL, color = NULL)
}
\arguments{
\item{p}{ggplot object}

\item{family}{font fammily}

\item{fontface}{font face}

\item{size}{font size}

\item{color}{font color}
}
\value{
TableGrob object
}
\description{
setting font for ggplot (axis text, label, title, etc.)
}
\examples{
library(grid)
library(ggplot2)
d <- data.frame(x=rnorm(10), y=rnorm(10), lab=LETTERS[1:10])
p <- ggplot(d, aes(x, y)) + geom_text(aes(label=lab), size=5)
g <- set_font(p, family="Times", fontface="italic", color='firebrick')
grid.draw(g)
}
\author{
guangchuang yu
}
