% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides2png.R
\name{slides2png}
\alias{slides2png}
\title{slides2png}
\usage{
slides2png(pdf, outfile = "slides.png", scale = 1, page = "all",
  add_page = TRUE)
}
\arguments{
\item{pdf}{pdf file}

\item{outfile}{output png file}

\item{scale}{scale of the width & height}

\item{page}{which pages to be converted}

\item{add_page}{whether add page number}
}
\value{
gtable
}
\description{
convert pdf slides to one single png figure
}
\author{
guangchuang yu
}
